Battle System Update: Introduce Battle strategies, support Double and Safari battles
This updates the battle system to separate the menu-handling logic from the decision-making logic.

It introduces a new concept called 'BattleStrategy' that are objects with a set of functions that can make decisions given a certain battle state.

For example, the battle strategy's `decide_turn(battle_state)` method is called for each round of a battle to determine the next action (fight, item, run, ...)

Likewise, there are methods for deciding whether an evolution should be allowed or aborted, which existing move should be replaced if a new one is about to be learned, what to do after the lead Pokémon faints, etc.

Bot modes can specify their preferred battle strategy in the `on_battle_started()` method. That way, a bot mode can have much greater control over how the battle system behaves without the user having to modify the config files all the time. (And, for example the Nugget Bridge mode could have its own 'try to lose' battle strategy.)

So far, I have mostly just aimed for feature parity and not introduced any other battle strategies.

As part of the battle system overhaul double battles are now supported, as are Safari encounters (in principle, the battle strategies still only switch to manual.)