Add a headless (no GUI) mode (#372)
This adds the command line argument `--headless` which allows running the bot without the GUI.

It's meant for people (i.e. me) that want to run the bot on a server system and remotely control it using the HTTP server.

Currently, it does not support creating or selecting profiles interactively, so a profile needs to exist and it needs to be specified when starting the bot.

I have also added Waitress as a dependency, which is a simple production-ready HTTP server. Using this avoids Flask's warning message ('This is a development server...') and also makes it less noisy in general (the access log is no longer written to stdout.)