Fix Discord ping for custom filter matches
The Discord message embed wasn't created properly if an encounter matched custom catch filters, leading to an error.

This fixes that, and also extracts some commonly repeated code in the Discord integration plugin.

Until now it was also possible for an encounter to be reported _twice_, if it happened to match both shiny/anti-shiny pings and a custom catch filter. I've made sure that an encounter is only reported once.